+++ neurokit/__init__.py
@@ -1,7 +1,7 @@
 """
 neurokit module.
 """
-__version__ = "0.1.9"
+__version__ = "0.1.92"
 
 from .miscellaneous import *
 from .statistics import *